# generated/job-api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ActorType": {
        "enum": [
          "OPERATOR",
          "SCHEDULER",
          "WORKER",
          "RECOVERY",
          "SYSTEM"
        ],
        "title": "ActorType",
        "type": "string"
      },
...
      "JobApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/JobApiError"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "trace_id",
          "generated_at",
          "error"
        ],
        "title": "JobApiErrorEnvelope",
        "type": "object"
      },
...
      "JobDeduplicatedEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDeduplicatedData"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "trace_id",
          "generated_at",
          "data"
        ],
        "title": "JobDeduplicatedEnvelope",
        "type": "object"
      },
...
      "JobEnqueuedEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobEnqueuedData"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "trace_id",
          "generated_at",
          "data"
        ],
        "title": "JobEnqueuedEnvelope",
        "type": "object"
      },
...
      "JobListEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobListData"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "trace_id",
          "generated_at",
          "data"
        ],
        "title": "JobListEnvelope",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "QUEUED",
          "CLAIMED",
          "RUNNING",
          "SUCCEEDED",
          "FAILED",
          "BLOCKED",
          "TIMED_OUT",
          "CANCEL_REQUESTED",
          "CANCELLED"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobType": {
        "enum": [
          "SNAPSHOT",
          "DEBATE",
          "REPLAY",
          "BACKTEST"
        ],
        "title": "JobType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "ServiceBearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Trading Agent Job Command API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Type"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "actor_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ActorType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Type"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Id"
            }
          },
          {
            "in": "query",
            "name": "requested_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested From"
            }
          },
          {
            "in": "query",
            "name": "requested_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Service authentication is required."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Request body exceeds the bounded service limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Request filters are invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "An unexpected internal error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Authentication or the durable repository is unavailable."
          }
        },
        "security": [
          {
            "ServiceBearerAuth": []
          }
        ],
        "summary": "List Jobs"
      },
      "post": {
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueJobBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDeduplicatedEnvelope"
                }
              }
            },
            "description": "An identical canonical job already exists."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnqueuedEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Service authentication is required."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "The idempotency key conflicts with another request."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Request body exceeds the bounded service limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Request data is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "An unexpected internal error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApiErrorEnvelope"
                }
              }
            },
            "description": "Authentication or the durable repository is unavailable."
          }
        },
        "security": [
          {
            "ServiceBearerAuth": []
          }
        ],
        "summary": "Create Job"
      }
    },
...
  }
}